Implemented all functions in GarageImpl, Covered all functions with tests

+++ GarageEmulator/src/main/java/ru/sbt/garage/entities/Garage.java
@@ -40,7 +40,7 @@ public interface Garage {
      * Complexity should be less than O(n)
      * @return removed car
      */
-    Car removeCar(int carId);
+    Car removeCar(long carId);
 
     /**
      * Complexity should be less than O(n)
